Pivotal ID # 185645993: Kotlin AWS S3 Client (#882)
https://www.pivotaltracker.com/story/show/185645993
- migrated code to kotlin sdk

+++ scheduler/tasks/pmc-processor-task/src/main/kotlin/ac/uk/ebi/pmc/persistence/docs/Documents.kt
@@ -40,7 +40,7 @@ data class SubmissionDocument(
     var status: SubmissionStatus,
     val sourceFile: String,
     val posInFile: Int,
-    val sourceTime: Int,
+    val sourceTime: Long,
     val files: List<ObjectId> = emptyList(),
     val updated: Instant = Instant.now(),
 ) {
